requirements.txt
robotframework
robotframework-requests
robotframework-faker
robotframework-jsonlibrary
requests